requirements.txt
python-dotenv
groq
tavily-python
fastapi